# doc/api.yaml
---
openapi: 3.0.0
paths:
  '/user':
    post:
      operationId: UserController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: The record has been successfully created.
        '400':
          description: Request body does not contain required fields
      tags:
        - User
    get:
      operationId: UserController_findAll
      parameters: []
      responses:
        '200':
          description: Get all records.
      tags:
        - User
  '/user/{id}':
    get:
      operationId: UserController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Get record by id if it exists.
        '400':
          description: Id is invalid (not uuid)
        '404':
          description: The record not found.
      tags:
        - User
    put:
      operationId: UserController_updatePassword
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/UpdatePasswordDto'
      responses:
        '200':
          description: The record successfully updated.
        '400':
          description: Id is invalid (not uuid)
        '403':
          description: Forbidden. Old Password is wrong
        '404':
          description: The record not found.
      tags:
        - User
    delete:
      operationId: UserController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The record successfully deleted.
        '400':
          description: Id is invalid (not uuid)
        '404':
          description: The record not found. Record with this id does not exist
      tags:
        - User
  '/track':
    post:
      operationId: TrackController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/CreateTrackDto'
      responses:
        '201':
          description: The record has been successfully created.
        '400':
          description: Request body does not contain required fields
      tags:
        - Track
    get:
      operationId: TrackController_findAll
      parameters: []
      responses:
        '200':
          description: Get all records.
      tags:
        - Track
  '/track/{id}':
    get:
      operationId: TrackController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Get record by id if it exists.
        '400':
          description: Id is invalid (not uuid)
        '404':
          description: The record not found.
      tags:
        - Track
    put:
      operationId: TrackController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/UpdateTrackDto'
      responses:
        '200':
          description: The record successfully updated.
        '400':
          description: Id is invalid (not uuid)
        '404':
          description: The record not found.
      tags:
        - Track
    delete:
      operationId: TrackController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The record successfully deleted.
        '400':
          description: Id is invalid (not uuid)
        '404':
          description: The record not found. Record with this id does not exist
      tags:
        - Track
  '/album':
    post:
      operationId: AlbumController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/CreateAlbumDto'
      responses:
        '201':
          description: The record has been successfully created.
        '400':
          description: Request body does not contain required fields
      tags:
        - Album
    get:
      operationId: AlbumController_findAll
      parameters: []
      responses:
        '200':
          description: Get all records.
      tags:
        - Album
  '/album/{id}':
    get:
      operationId: AlbumController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Get record by id if it exists.
        '400':
          description: Id is invalid (not uuid)
        '404':
          description: The record not found.
      tags:
        - Album
    put:
      operationId: AlbumController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/UpdateAlbumDto'
      responses:
        '200':
          description: The record successfully updated.
        '400':
          description: Id is invalid (not uuid)
        '404':
          description: The record not found.
      tags:
        - Album
    delete:
      operationId: AlbumController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The record successfully deleted.
        '400':
          description: Id is invalid (not uuid)
        '404':
          description: The record not found. Record with this id does not exist
      tags:
        - Album
  '/artist':
    post:
      operationId: ArtistController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/CreateArtistDto'
      responses:
        '201':
          description: The record has been successfully created.
        '400':
          description: Request body does not contain required fields
      tags:
        - Artist
    get:
      operationId: ArtistController_findAll
      parameters: []
      responses:
        '200':
          description: Get all records.
      tags:
        - Artist
  '/artist/{id}':
    get:
      operationId: ArtistController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Get record by id if it exists.
        '400':
          description: Id is invalid (not uuid)
        '404':
          description: The record not found.
      tags:
        - Artist
    put:
      operationId: ArtistController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/UpdateArtistDto'
      responses:
        '200':
          description: The record successfully updated.
        '400':
          description: Id is invalid (not uuid)
        '404':
          description: The record not found.
      tags:
        - Artist
    delete:
      operationId: ArtistController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The record successfully deleted.
        '400':
          description: Id is invalid (not uuid)
        '404':
          description: The record not found. Record with this id does not exist
      tags:
        - Artist
  '/favs':
    get:
      operationId: FavsController_findAll
      parameters: []
      responses:
        '200':
          description:
            Get all records (all favorite records (not their ids), split
            by entity type).
      tags:
        - Favs
  '/favs/track/{id}':
    post:
      operationId: FavsController_addTrack
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: The record has been successfully added to favorites.
        '400':
          description: Id is invalid (not uuid)
        '422':
          description: Record with this id does not exist.
      tags:
        - Favs
    delete:
      operationId: FavsController_removeTrack
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The record successfully removed from favorites.
        '400':
          description: Id is invalid (not uuid)
        '422':
          description: Corresponding record is not in favorites.
      tags:
        - Favs
  '/favs/album/{id}':
    post:
      operationId: FavsController_addAlbum
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: The record has been successfully added to favorites.
        '400':
          description: Id is invalid (not uuid)
        '422':
          description: Record with this id does not exist.
      tags:
        - Favs
    delete:
      operationId: FavsController_removeAlbum
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The record successfully removed from favorites.
        '400':
          description: Id is invalid (not uuid)
        '422':
          description: Corresponding record is not in favorites.
      tags:
        - Favs
  '/favs/artist/{id}':
    post:
      operationId: FavsController_addArtist
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: The record has been successfully added to favorites.
        '400':
          description: Id is invalid (not uuid)
        '422':
          description: Record with this id does not exist.
      tags:
        - Favs
    delete:
      operationId: FavsController_removeArtist
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The record successfully removed from favorites.
        '400':
          description: Id is invalid (not uuid)
        '422':
          description: Corresponding record is not in favorites.
      tags:
        - Favs
info:
  title: Home Library Service
  description: Home Library Service API description
  version: '1.0'
  contact: {}
tags: []
servers: []
components:
  schemas:
    CreateUserDto:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
      required:
        - login
        - password
    UpdatePasswordDto:
      type: object
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
      required:
        - oldPassword
        - newPassword
    CreateTrackDto:
      type: object
      properties:
        name:
          type: string
        artistId:
          type: string
        albumId:
          type: string
        duration:
          type: number
      required:
        - name
        - artistId
        - albumId
        - duration
    UpdateTrackDto:
      type: object
      properties:
        name:
          type: string
        artistId:
          type: string
        albumId:
          type: string
        duration:
          type: number
      required:
        - name
        - artistId
        - albumId
        - duration
    CreateAlbumDto:
      type: object
      properties:
        name:
          type: string
        year:
          type: number
        artistId:
          type: string
          nullable: true
      required:
        - name
        - year
        - artistId
    UpdateAlbumDto:
      type: object
      properties:
        name:
          type: string
        year:
          type: number
        artistId:
          type: string
          nullable: true
      required:
        - name
        - year
        - artistId
    CreateArtistDto:
      type: object
      properties:
        name:
          type: string
        grammy:
          type: boolean
      required:
        - name
        - grammy
    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
        grammy:
          type: boolean
      required:
        - name
        - grammy